+++ Source/NETworkManager/Models/Network/SubnetInfo.cs
@@ -6,11 +6,11 @@ namespace NETworkManager.Models.Network
     {
         public IPAddress NetworkAddress { get; set; }
         public IPAddress Broadcast { get; set; }
-        public int IPAddresses { get; set; }
+        public long IPAddresses { get; set; }
         public IPAddress Subnetmask { get; set; }
         public int CIDR { get; set; }
         public IPAddress HostFirstIP { get; set; }
         public IPAddress HostLastIP { get; set; }
-        public int Hosts { get; set; }
+        public long Hosts { get; set; }
     }
 }
